pyproject.toml
[tool.poetry]
name = "tradestation-api-python"
version = "1.3.0"
description = "A Python wrapper for the TradeStation WebAPI, providing easy access to brokerage, order execution, and market data services."
authors = ["mxcoppell"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mxcoppell/tradestation-api-python"
packages = [{ include = "tradestation", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.12.14"
python-dotenv = "^1.0.0"
requests = "^2.32.4"
urllib3 = "^2.6.0"
pydantic = "^2.10.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
black = "^24.3.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.flake8]
max-line-length = 100
extend-ignore = "E203"
